+++ app/src/main/res/layout/activity_review.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<RelativeLayout
+<androidx.constraintlayout.widget.ConstraintLayout
     tools:context=".ReviewActivity"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
@@ -14,11 +14,9 @@
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginEnd="145dp"
-        android:layout_marginRight="145dp"
-        android:layout_marginBottom="110dp"
         android:text="작성"
         app:layout_constraintEnd_toEndOf="@+id/et_contents"
+        app:layout_constraintHorizontal_bias="0.502"
         app:layout_constraintStart_toStartOf="@+id/et_contents"
         app:layout_constraintTop_toBottomOf="@+id/et_contents" />
 
@@ -29,9 +27,9 @@
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
         android:layout_alignParentBottom="true"
+        android:layout_marginTop="20dp"
         android:layout_marginEnd="19dp"
         android:layout_marginRight="19dp"
-        android:layout_marginBottom="198dp"
         android:ems="10"
         android:hint="내용"
         android:inputType="text"
@@ -46,15 +44,12 @@
         android:layout_height="wrap_content"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
-        android:layout_marginStart="8dp"
-        android:layout_marginLeft="8dp"
+        android:layout_marginStart="100dp"
+        android:layout_marginLeft="100dp"
         android:layout_marginTop="50dp"
-        android:layout_marginEnd="84dp"
-        android:layout_marginRight="84dp"
         android:ems="10"
         android:hint="제목"
         android:inputType="text"
-        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -65,9 +60,7 @@
         android:layout_alignParentTop="true"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
-        android:layout_marginTop="179dp"
-        android:layout_marginEnd="84dp"
-        android:layout_marginRight="84dp"
+        android:layout_marginTop="10dp"
         android:isIndicator="false"
         android:max="5"
         android:numStars="5"
@@ -77,4 +70,4 @@
         app:layout_constraintStart_toStartOf="@+id/et_title"
         app:layout_constraintTop_toBottomOf="@+id/et_title" />
 
-</RelativeLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
